Extract gRPC common stuff
