+++ extensions/grpc-common/runtime/src/main/resources/META-INF/quarkus-extension.yaml
@@ -1,0 +1,11 @@
+---
+name: "gRPC Common"
+metadata:
+  keywords:
+    - "gRPC"
+  categories:
+    - "web"
+    - "serialization"
+    - "reactive"
+  status: "experimental"
+  unlisted: "true"
