+++ extensions/grpc/deployment/src/main/java/io/quarkus/grpc/deployment/GrpcServerProcessor.java
@@ -10,10 +10,7 @@ import org.eclipse.microprofile.config.ConfigProvider;
 import org.jboss.jandex.ClassInfo;
 import org.jboss.logging.Logger;
 
-import io.grpc.internal.DnsNameResolverProvider;
-import io.grpc.internal.PickFirstLoadBalancerProvider;
 import io.grpc.internal.ServerImpl;
-import io.grpc.netty.NettyChannelProvider;
 import io.quarkus.arc.deployment.AdditionalBeanBuildItem;
 import io.quarkus.arc.processor.DotNames;
 import io.quarkus.deployment.IsDevelopment;
@@ -28,8 +25,6 @@ import io.quarkus.deployment.builditem.ExtensionSslNativeSupportBuildItem;
 import io.quarkus.deployment.builditem.FeatureBuildItem;
 import io.quarkus.deployment.builditem.ServiceStartBuildItem;
 import io.quarkus.deployment.builditem.ShutdownContextBuildItem;
-import io.quarkus.deployment.builditem.nativeimage.NativeImageConfigBuildItem;
-import io.quarkus.deployment.builditem.nativeimage.ReflectiveClassBuildItem;
 import io.quarkus.grpc.deployment.devmode.FieldDefinalizingVisitor;
 import io.quarkus.grpc.runtime.GrpcContainer;
 import io.quarkus.grpc.runtime.GrpcServerRecorder;
@@ -99,42 +94,6 @@ public class GrpcServerProcessor {
     }
 
     @BuildStep
-    public void configureNativeExecutable(CombinedIndexBuildItem combinedIndex,
-            BuildProducer<ReflectiveClassBuildItem> reflectiveClass,
-            BuildProducer<ExtensionSslNativeSupportBuildItem> extensionSslNativeSupport) {
-
-        // we force the usage of the reflection invoker.
-        Collection<ClassInfo> messages = combinedIndex.getIndex()
-                .getAllKnownSubclasses(GrpcDotNames.GENERATED_MESSAGE_V3);
-        for (ClassInfo message : messages) {
-            reflectiveClass.produce(new ReflectiveClassBuildItem(true, true, true, message.name().toString()));
-        }
-        Collection<ClassInfo> builders = combinedIndex.getIndex().getAllKnownSubclasses(GrpcDotNames.MESSAGE_BUILDER);
-        for (ClassInfo builder : builders) {
-            reflectiveClass.produce(new ReflectiveClassBuildItem(true, true, true, builder.name().toString()));
-        }
-
-        Collection<ClassInfo> lbs = combinedIndex.getIndex().getAllKnownSubclasses(GrpcDotNames.LOAD_BALANCER_PROVIDER);
-        for (ClassInfo lb : lbs) {
-            reflectiveClass.produce(new ReflectiveClassBuildItem(true, true, false, lb.name().toString()));
-        }
-
-        Collection<ClassInfo> nrs = combinedIndex.getIndex().getAllKnownSubclasses(GrpcDotNames.NAME_RESOLVER_PROVIDER);
-        for (ClassInfo nr : nrs) {
-            reflectiveClass.produce(new ReflectiveClassBuildItem(true, true, false, nr.name().toString()));
-        }
-
-        // Built-In providers:
-        reflectiveClass.produce(new ReflectiveClassBuildItem(true, true, false, DnsNameResolverProvider.class));
-        reflectiveClass.produce(new ReflectiveClassBuildItem(true, true, false, PickFirstLoadBalancerProvider.class));
-        reflectiveClass.produce(new ReflectiveClassBuildItem(true, true, false,
-                "io.grpc.util.SecretRoundRobinLoadBalancerProvider$Provider"));
-        reflectiveClass.produce(new ReflectiveClassBuildItem(true, true, false, NettyChannelProvider.class));
-
-        extensionSslNativeSupport.produce(new ExtensionSslNativeSupportBuildItem(GRPC_SERVER));
-    }
-
-    @BuildStep
     HealthBuildItem addHealthChecks(GrpcServerBuildTimeConfig config,
             List<BindableServiceBuildItem> bindables,
             BuildProducer<AdditionalBeanBuildItem> beans) {
@@ -157,17 +116,7 @@ public class GrpcServerProcessor {
     }
 
     @BuildStep
-    NativeImageConfigBuildItem nativeImageConfiguration() {
-        NativeImageConfigBuildItem.Builder builder = NativeImageConfigBuildItem.builder()
-                .addRuntimeInitializedClass("io.grpc.netty.Utils$ByteBufAllocatorPreferDirectHolder")
-                .addRuntimeInitializedClass("io.grpc.netty.Utils$ByteBufAllocatorPreferHeapHolder")
-                // substitutions are runtime-only, Utils have to be substituted until we cannot use EPoll
-                // in native. NettyServerBuilder and NettyChannelBuilder would "bring in" Utils in build time
-                // if they were not marked as runtime initialized:
-                .addRuntimeInitializedClass("io.grpc.netty.Utils")
-                .addRuntimeInitializedClass("io.grpc.netty.NettyServerBuilder")
-                .addRuntimeInitializedClass("io.grpc.netty.NettyChannelBuilder");
-        return builder.build();
+    ExtensionSslNativeSupportBuildItem extensionSslNativeSupport() {
+        return new ExtensionSslNativeSupportBuildItem(GRPC_SERVER);
     }
-
 }
